Case One_ Connecting ADK to GCP/requirements.txt
# Core ADK and GenAI SDKs
google-adk==1.7.0
google-genai==1.24.0

# Google Cloud APIs
google-cloud-bigquery==3.31.0
google-cloud-dlp==3.28.0

# Logging
loguru==0.7.3

# Web Framework (if using FastAPI-based interface or server)
fastapi==0.116.1

# Optional: typing support for Python <3.9
typing-extensions>=4.0.0

fastmcp
